feat(voyage): le Plan de vol emménage dans l'arbre
Deuxième vue à vivre dans la racine unique, après la Tournée. Elle était le candidat naturel :
l'ADR-004 en fait une CONCLUSION, où « rien n'est actionnable — pas un bouton de vente, pas un ✕,
pas un champ ». Aucune action à recâbler, donc.

`planHypotheses`, `planData` et le rendu passent dans `vues/plan-vue.tsx` ; `renderPlan()` disparaît
et `refresh()` ne nomme plus cette vue. `jambeChargee` rejoint `voyage-donnees.ts`, dont il ne
lisait que le registre.

DEUX PORTAILS POUR UNE VUE, et ce n'est pas un détail : `#planHead` et `#planBody` sont séparés dans
`index.html` par la carte du parcours. Un élément à écouteurs directs se déménage en FRÈRE, jamais
en enfant d'un conteneur réécrit (leçon de #24, rappelée par l'ADR-004) — la structure est donc
reproduite telle quelle.

Un commentaire est devenu faux au passage et il est corrigé plutôt que laissé : l'écouteur de
« ⧉ Copier » est posé sur `#planHead` parce que son contenu est réécrit à chaque rendu. C'était vrai
à cause de `renderPlan` ; c'est vrai maintenant à cause de l'arbre. Même règle, autre raison.

Les 19 tests de `plan.pw.mjs` passent sans avoir été touchés.

`app.js` : 2 942 → 2 869 lignes. 35 `peindre()` restants sur les 41 du départ.

Contribue à #96.
